clean up vk directory
