+++ sdk/engine/source/graphics/vk/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
         Initializers.cpp
+        QueueFamily.cpp
+        RenderPass.cpp
         Swapchain.cpp
-        VkTypes.cpp
 )
