+++ sdk/engine/source/graphics/vk/Common.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace nc::graphics
+{
+enum class ShaderStage
+{
+    Vertex,
+    Pixel
+};
+} // namespace nc::graphics
