+++ sdk/engine/source/graphics/vk/RenderPass.h
@@ -6,38 +6,13 @@
 #include "vulkan/vk_mem_alloc.hpp"
 #include <array>
 #include <memory>
-#include <optional>
 #include <span>
 #include <vector>
 
+/** @todo As more stuff gets pulled into here, consider splitting this up into multiple files. */
+
 namespace nc::graphics
 {
-enum class QueueFamilyType : uint8_t
-{
-    GraphicsFamily,
-    PresentFamily
-};
-
-class QueueFamilyIndices
-{
-    public:
-        QueueFamilyIndices(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface);
-        bool IsComplete() const;
-        bool IsSeparatePresentQueue() const;
-        uint32_t GetQueueFamilyIndex(QueueFamilyType type) const;
-
-    private:
-        std::optional<uint32_t> m_graphicsFamily;
-        std::optional<uint32_t> m_presentFamily;
-        bool m_isSeparatePresentQueue;
-};
-
-enum class ShaderStage : uint8_t
-{
-    Vertex,
-    Pixel
-};
-
 using ClearValueFlags_t = uint8_t;
 
 struct ClearValueFlags
